added test

+++ src/main/java/task5/atm_task/addMoney_utils/AddMoneyBehavior.java
@@ -2,10 +2,12 @@ package task5.atm_task.addMoney_utils;
 
 import com.sun.istack.internal.NotNull;
 
+import java.math.BigDecimal;
+
 /**
  *  Part of implementation of pattern Strategy
  */
 
 public interface AddMoneyBehavior {
-    double addMoney(@NotNull double amount);
+    BigDecimal addMoney(@NotNull double amount);
 }
