+++ src/main/java/task5/atm_task/wasteMoney_utils/WasteMoneyBehavior.java
@@ -2,10 +2,12 @@ package task5.atm_task.wasteMoney_utils;
 
 import com.sun.istack.internal.NotNull;
 
+import java.math.BigDecimal;
+
 /**
  * Part of implementation of pattern Strategy
  */
 
 public interface WasteMoneyBehavior {
-    double removeMoney(@NotNull double amount);
+    BigDecimal removeMoney(@NotNull double amount);
 }
